+++ pages/houtbouw/gemeentelijke-omgevingsvisie.js
@@ -2,7 +2,7 @@ import Link from 'next/link';
 import Image from 'next/image';
 import Layout from '../../components/layout';
 import IcontWood from '../../public/icons/houtbouwIconBg.svg';
-import RTooltip from '../../components/r_ladder_tooltip';
+import RTooltip from '../../components/r-ladder-tooltip';
 import JHTooltip from '../../components/juridische-houdbaarheid-tooltip';
 import JITooltip from '../../components/juridische-invloed-tooltip';
